requirements.txt
glfw==1.7.0
numpy==1.15.3
PyOpenGL==3.1.0
